requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0

SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

Flask-Migrate==4.0.7
flask
pandas
scikit-learn
matplotlib
joblib